Expressions for BCs and ICs are now symbolic
By doing this and using the Julia Symbolics package it is
possible to do symbolic manipulation of these expressions.
The immediate usage is to compute symbolic time derivatives
of DBCs so that velocities and accelerations at the
corresponding DOFs are exact.

Once the symbolic expresions and derivatives are obtained,
they can be evaluated by substituting the variables for
desired values. This proved tricky, in that Julia sometimes
gave an expression and sometimes a Float64 as a result
without being clear why. So there is a hackish extract_value
function that deals with this ambiguity.

This symbolic ability is intriguing, and coud be used to
do more complicated things, such as computation of stress
and moduli for constitutive models, integration for
transfer operators, etc.