Remove static array types.
The rationale to introduce them was efficiency and speed,
as the storage is preallocated for static arrays. This
was to mirror the MiniTensor library in Trilinos.

Nevertheless, the static array types do not inherit all the methods
applicable to general vectors and matrices, among them the
logarithmic and exponential maps, which are essential for
finite-deformation constitutive models. Also, by removing the
static arrays, there is no significant performance difference
and the code is simplified a bit.