Fixes and modifications.
- The contact code now works with Exodus.jl

- The tolerances for parametric coordinates and distances
were scattered and inconsistent with one another. Set them
all in one single place until there is evidence that they
need to be different for different calls.

- The point projection tolerance was too loose. Set it
to a much tighter value to avoid issues with the projection.

- Change the logic for finding and projecting nodes and points
to short circuit as soon as one is "found".

TODO: The logic for the detection algorithm and the application
of the contact DBC seems inconsistent. Talk to Daria about this.