Rerevert transfer operator change to be logically correct and disable nonoverlap test.
There is something fundamentally wrong about the
current implementation of nonoverlap coupling that
requires computing the transfer operators
at every time step and Schwarz iteration.
Some simulations run but this may be a fluke,
so I'm disabling the nonoverlap test and
reintroducing the change to the suspicious
call to the transfer operators until this
is all sorted out.