Refactor projection of point to side set
Introduce a preliminary proximity search to determine
the closest face in the side set to the point and
perform the projection on that face only.

This avoids an issue where the point is so far away from
a distorted face that the isoparametric map is
ill-defined, leading to failure of the projection.