Consolidate evaluators in model
Consolidate all evaluators for the different
time integrators in model.jl into a single
one to avoid much dupplicated code.

We also noticed by profiling that the code for
inclined DBCs introduces a performace decrease.
We need to refactor this to reduce the performace
impact.

We also need to find ways to improve performance
to run larger meshes in reasonable times.